/registration and /clients apge updated
On registration page user can save client, after which, he is redirected to the  /clients and there will be loaded random data and showed at the bottom of the page.
On /clients page clents in the table are loading dinamicly. At the bottom is showed the last str, which was generated from the last client save on the /client.